+++ Cargo.toml
@@ -8,7 +8,7 @@ anyhow = "1.0.97"
 ascii_table = "4.0.7"
 bincode = { version = "2.0.1", features = ["serde"] }
 chrono = "0.4.41"
-clap = { version = "4.5.41", features = ["derive"] }
+clap = { version = "4.5.41", features = ["derive", "env"] }
 dirs = "6.0.0"
 itertools = "0.14.0"
 once_cell = "1.21.3"
